move monotonically_increasing_id
