+++ daft/daft/__init__.pyi
@@ -1197,7 +1197,6 @@ def minhash(
     seed: int = 1,
     hash_function: Literal["murmurhash3", "xxhash", "sha1"] = "murmurhash3",
 ) -> PyExpr: ...
-def monotonically_increasing_id() -> PyExpr: ...
 def row_number() -> PyExpr: ...
 def rank() -> PyExpr: ...
 def dense_rank() -> PyExpr: ...
