--- src/daft-functions/src/python/sequence.rs
@@ -1,8 +1,0 @@
-use daft_dsl::python::PyExpr;
-use pyo3::{pyfunction, PyResult};
-
-simple_python_wrapper!(
-    monotonically_increasing_id,
-    crate::sequence::monotonically_increasing_id,
-    []
-);
