--- src/daft-functions/src/sequence/mod.rs
@@ -1,3 +1,0 @@
-pub mod monotonically_increasing_id;
-
-pub use monotonically_increasing_id::monotonically_increasing_id;
