+++ src/daft-logical-plan/src/optimization/rules/detect_monotonic_id.rs
@@ -116,7 +116,7 @@ impl OptimizerRule for DetectMonotonicId {
 mod tests {
     use common_error::DaftResult;
     use common_scan_info::Pushdowns;
-    use daft_functions::sequence::monotonically_increasing_id::monotonically_increasing_id;
+    use daft_functions::monotonically_increasing_id::monotonically_increasing_id;
     use daft_schema::{dtype::DataType, field::Field};
 
     use crate::{
